Website/requirements.txt
flask
flask-mysql-connector
mysql-connector